Makefile overhaul

+++ rtsTrivial/rts.c
@@ -31,7 +31,7 @@ void putStr(LIST xs) {
       putchar('\n');
       return;
     default: 
-      x = (Char)((CONS)xs)->a;
+      x = (Char)(Int)((CONS)xs)->a;
       xs = ((CONS)xs)->b;
       putchar(x);
       break;
@@ -44,11 +44,11 @@ LIST getStr(char *p) {
         return (LIST)_NIL;
     CONS n0; NEW(CONS, n0, sizeof(struct CONS));
     CONS n = n0;
-    n->a = (POLY)*p++;
+    n->a = (POLY)(Int)*p++;
     while (*p) {
         NEW(LIST, n->b, sizeof(struct CONS));
         n = (CONS)n->b;
-        n->a = (POLY)*p++;
+        n->a = (POLY)(Int)*p++;
     }
     n->b = (LIST)_NIL;
     return (LIST)n0;
